fix_naive.cpp

+++ data_structure/static_range_sum_with_upper_bound/sol/naive.cpp
@@ -19,11 +19,11 @@ int main(){
             }
             printf("%d\n",ans);
         }else{
-            long ans = 0;
+            long long ans = 0LL;
             for(int i = l;i < r;i++){
                 if(a[i] <= x) ans += a[i];
             }
-            printf("%ld\n",ans);
+            printf("%lld\n",ans);
         }
     }
